refactor:컴포넌트를 common과 page로 구분
