+++ src/components/page/OrderConfirmationPage/CheckedItem/style.ts
