+++ src/components/page/OrderConfirmationPage/CheckedItemContainer/style.ts
