+++ src/components/page/OrderConfirmationPage/ConfirmationContainer/style.ts
