+++ src/components/page/OrderConfirmationPage/DeliveryInfo/style.ts
