+++ src/components/page/PaymentPage/style.ts
