+++ src/components/page/ShoppingCartPage/CartItem/style.ts
